STORE-2337 Improvement: Add the ability to add multiple attributes in the user submission tools.
